Updated openstack/openstack Project: openstack/python-cinderclient  d5c5bacb67a53086efcda4b8f5f675056ba8b2d8
Mask passwords in client debug output

This change looks for the use of 'password' in the data that
is sent and uses mask_password() to remove the actual password
text.  This change will prevent debug output that is being
saved from saving passwords.

A test case is added to verify that password output is being removed.

Change-Id: I93bde838ea21101df08c0e824d9f9457ed2ad077
Closes-Bug: 1341735