Updated openstack/openstack Project: openstack/nova  eb2a071ce4ef9748958f315a1fc6e8c26f1d3665
Parse unicode cpu_info as json before using it

If the extra specs are some of this:
- (capabilities:cpu_info:vendor, Intel)
- (capabilities:cpu_info:topology:cores, 2)
- (capabilities:cpu_info:features, <in> rdtscp)

It doesn't work because the cpu_info is loaded as unicode in HostState.
The patch fix these cases by parsing the cpu_info to dict.
Update: use six.string_types instead of unicode.

Change-Id: Ieabf95b1e3e5893ced16c0571c29b13b252df5d1
Closes-Bug: #1331176
Co-Authored-By: Pawel Koniszewski <pawel.koniszewski@intel.com>