Updated openstack/openstack Project: openstack/requirements  ae6e876ea820941fb994dda470436e7605642484
Unbind version of pyzmq

Currently pyzmq supports 2 branches with some API differences:
2.* - legacy branch
14.* - current branch

Due to historic reasons we bind pyzmq to an old API version.
In devstack we install latest version of pyzmq and not 2.* branch.
Cause of this I don't see any reason for us to bind pyzmq to such an old version.

Updating to latest version of pyzmq will allow us to implement zmq
driver in oslo.messaging that won't depend on eventlet and will be Py3
compatible.

Change-Id: I009e61303c01324de29ec78202621f50420d4d7f