Updated openstack/openstack Project: openstack/keystone  89d47a0542753c00272da906ce6f3fdcba988b80
Don't override tox envdir for pep8 and cover jobs

infra's run-pep8 and run-coverage jobs assume the envdir for pep8 and
cover jobs are not overwritten, so it can call out to .tox/pep8/bin/pip
and .tox/coverage/bin/pip respectively.

This bug is causing keystone pep8 and coverage jobs to occasionally fail
in the gate.

If the motivation behind overriding the envdir jobs to use venv, one can
use devpi and/or wheels to speed up the process of building a venv.

Change-Id: I32cd46152caa73e1308fdd216ba016c80a5a9b4f
Closes-Bug: #1352134