Updated openstack/openstack Project: openstack/nova  6570f45c1daf1741bf2cc1ccc3f85f4aa69c9805
Make spawn_n() stub properly ignore errors in the child thread work

When we call spawn_n() normally, we fork off a thread that can run or
die on its own, without affecting the parent. In unit tests, we stub
this out to be a synchronous call, but we allow any exceptions that
occur in that work to bubble up to the caller. This is not normal
behavior and thus we should discard any such exceptions in order to
mimic actual behavior of a child thread.

Change-Id: I35ab21e9525aa76cced797436daa0b99a4fa99f2
Related-bug: #1349147