Updated openstack/openstack Project: openstack/nova  5278181e09d4f4184ca9a115a1dbe460bbcd6e08
Add policy on how patches and reviews go hand in hand

There is currently a misconception about how a patch gets merged:
Author proposes a patch, review team reviews it. Us, and Them. Instead
it should be a 'We.'
If the review bandwidth doesn't grow at the same rate as the patch
proposal rate then we end up with a massive backlog where people are
pinging cores for reviews.

Record a recommended policy of asking patch authors to offset the review
resources spent on their patch by reviewing other patches.

Change-Id: I50653f8ea58240fa664519f544ce16b2adf39006