Updated openstack/openstack Project: openstack/sahara  27704ba3303bcfd8c5034d385ad2d957b035e374
Fixed a ValueError on provisioning cluster

If the machine is slow, the remote command may output an empty string.
Converting an empty string to an integer results a ValueError, which
caused a failure. This patch attempted to fix the error by adding a
check for the outputted string before converting to integer.

Closes-Bug: 1351624
Change-Id: I38aadaa8c6b3ae9c7f93f90d5144ac8de7aa54d4