Updated openstack/openstack Project: openstack/nova  82c5ee548603ef2e5219265f84244b922796dc79
Handle MacAddressInUseClient exception from Neutron when creating port

This adds exception handling for the new MacAddressInUseClient exception
from python-neutronclient-2.3.6 when trying to create a port with a
specific MAC address.

The PortInUse exception is used so that the build will be rescheduled if
this error is encountered.

Also, the servers API already handles PortInUse and translates it to an
HTTPConflict response.

Closes-Bug: #1347778

Change-Id: I929b5f9d79bcb7e60e8b3482e5ee1ca2fababed1