Updated openstack/openstack Project: openstack/nova  f48827dcfa8d99948ead79626be53c3031a53496
Update scheduler after instance delete

Previously deleting an instance would update the scheduler resources
fairly quickly. There is now a delay when deleting an instance until the
scheduler makes the resources available again. This appears to be due to
the fact that the delete code path used to call resource tracker to
update the compute_node record but this no longer happens.

Closes-Bug: 1336080
Change-Id: I837b70ce06843ff83481d4985c9c08b74d20cab0