Updated openstack/openstack Project: openstack-infra/devstack-gate  1495d8d28f9020e814ed8c2e35eb8f4e46d245fd
Disable cinder with ironic

Cinder doesn't currently work with ironic as volumes cannot yet
be attached to baremetal nodes. This is something that may be
available in the K series. For now, though, so that we can work
on enabling the tempest API and scenario tests with ironic, we
need to disable cinder.

Change-Id: I4e2afc232986f70263425deec4b335a1a5b8110f