Updated openstack/openstack Project: openstack/horizon  bff1ebd1497540b6f28fe767ad376e3389ce6868
TEMPLATE_DIRS must be a tuple

With Django 1.7, TEMPLATE_DIRS must be a tuple. The
definition in horizon/test/settings.py must therefore
have a leading comma. Note that other definitions of
TEMPLATE_DIRS are already like this in Horizon, so
this just fixes the one instance which is wrong.

Change-Id: I4d9aa9e716e4cf6c2b56fd25ab8fde9187fc5108