Updated openstack/openstack Project: openstack/horizon  81d14c715ef322d591969d4dd93c3203b87b82d8
Change note about the removal of variable SESSION_SERIALIZER

The comment should be changed to reflect the results of the
proposed removal of the variable during the J-cycle. More details
are in the comments of the change request I625f5f4a9f128ff31aee58f2a79ee88c91dc1f56.

Related to bug #1349463.

Change-Id: I625f5f4a9f128ff31aee58f2a79ee88c91dc1f56