Updated openstack/openstack Project: openstack/cinder  07534bef77704e03f73e34adea594f06c09adcda
Fix error log level in restore-backup routine

When doing backup-restore operation, it needs volume's size is not less
than backup's size.
But in backup-restore routine, when volume's size is bigger than backup's
size, it logs the the info of volume's size and backup's size to warn
level, which I think it should log the info to info level.
This bug fix it.

Change-Id: Ic7210e3b018053be7a6f567cd0a7b20d51d403b8
Closes-Bug: #1338481