Updated openstack/openstack Project: openstack-infra/devstack-gate  a0a61f9cf1b89e9c5ccd624b31c72f138cfbba87
Revert "Disable cinder with ironic"

This reverts commit 034bbcc51d30a77eab7da8b4f7550b4a9d8a2552.

The ironic gate tests that made use of the libvirt driver (for
exercising the ironic API only) broke due to this change since
they try to test server rescue. We can't set the rescue feature
flag to False without affecting other tests using libvirt.

We are working on merging the ironic gate tests so that they all
will use the ironic driver. Once that is done, we can disable
cinder again and set the rescue feature flag properly.

Closes-Bug: #1353027

Change-Id: I059f06f4289ea332b49b3963e0d544e254298ac8