Updated openstack/openstack Project: openstack/horizon  1279d4a7c071c384e8b4e7bd62109db832800273
Support provider network extension when creating network

This adds the ability for admins to set the provider network
settings when creating a network. This includes the network type,
physical network, and segmentation ID. These options will only be
available if the provider network neutron extension is supported.

Change-Id: I6108504a92a32f067cf151ec103171a7bbb601df
Implements: blueprint quantum-network-provider-types
Closes-Bug: 1223360