Updated openstack/openstack Project: openstack/neutron  d8c70257b03c9877467d9efc771750b4209d7c8c
Simple refactor to stop passing around an unused parameter

There is no functional change in this patch.  However, this refactor
allows another patch [1] to make a simpler change to the code base.

This patch includes localizing the initialization of "internal_cidrs"
to a later spot in the "process_router" method.  This has the
advantage that it will be initialized under the same condition that it
will be used.

[1] https://review.openstack.org/#/c/109776

Change-Id: I3ca993d0a33b7c925526099e97999c9d9f76efc2