Updated openstack/openstack Project: openstack/horizon  354d89bc7a2746d02746f780406aabff79728ac3
template to rely on the the "id" attribute

The exception fixed is the following:
NoReverseMatch: Reverse for 'detail' with arguments '('',)' and keyword
arguments '{}' not found. 1 pattern(s) tried:
[u'project/volumes/(?P<volume_id>[^/]+)/$']

The test data doesn't include a volumeId attribute to volumes but they
contain an id attribute.

Here I modify the template to rely on the the "id" attribute but maybe
real data has a volumeId attribute and it's the test data in
openstack_dashboard/test/test_data/nova_data.py that needs to be
modified instead.

Change-Id: Ia561c2db1c59e51fb669e1ad6636d9b3bc9f7676