Updated openstack/openstack Project: openstack/python-cinderclient  c01529fa538d28af3a9f0f7cdc1a96a9df5ed221
Add missing all-tenants option to transfer-list

cinder transfer-list is missing the --all-tenants option, which makes
it inconsistent with 'cinder list'. Without this fix, when running as
admin horizon will need to make multiple queries to cinder (one for
each tenant present in the volume list) in order to display the transfers
that exist for the volumes being shown.

Change-Id: Ic4d72a54c450876ff24c64efd83678062c93a24e
Closes-bug: 1351084