Updated openstack/openstack Project: openstack/trove  1fd4c4508dd53ae5bea22ce2380d917921aa09cd
guestagent/test_volume.py leaves a file in /tmp

Since the test mocks() execute and then calls write_to_fstab(), the
file created in /tmp by write_to_fstab() stays around. This is because
write_to_fstab() uses execute to remove the file. Replaced it with an
os.remove() and fixed the test which was counting two calls to
execute().

I'm fixing it because testing for another bug involves running tox
over and over again and leaving these files in /tmp is just a bit
messy.

Change-Id: Ia9e01049687644abd0671a658dffbbdcb4fba4bd
Closes-Bug: #1353064