Updated openstack/openstack Project: openstack/nova  3ecbde03c1c33b09f295295e7de698b48a6f6f18
API: Enable support for tenant option in nova absolute-limits

When querying for the absolute limits of a specific tenant,
the tenant option is ignored. There are no attempts to extract
the tenant from the request. Instead, nova uses context.project_id
as the project_id in QUOTAS.get_project_quotas. The following
patch extracts the tenant_id from the request (if any) and passes
that to QUOTAS.get_project_quotas to obtain the proper quota.

Change-Id: If5f91de020ed8a40fa04fc001c7c4c92681f4ad1
Closes-Bug: #1334278