Updated openstack/openstack Project: openstack/ceilometer  c656ae1bff6e0f001c985a5f1d775576bcea8001
[HBase] Improve uniqueness for row in meter table

Currently, row in meter table consists of reversed timestamp,
meter and an md5 of user+resource+project for purposes of
uniqueness. But md5 of user+resource+project may be the same
for different meters and only "rts" provides the difference.
But in the tests or under the high load different meters
from the same user, resource and project may have the same
timestamp.
In this patchset row contains a message signature instead of
md5 of user+resource+project, because message signature has
dependecies from all fields of sample, it provides more uniqueness.

Change-Id: Ic767b257176e05c0b35664968e7f529f507eab11