Updated openstack/openstack Project: openstack/nova  05ab558cf7fcba01ab99b49f2c45cb85e0d741d2
Use hypervisor hostname for compute trust level

In XenAPI, service hostname and compute node hostname is different
because the Nova compute service may run in a separated VM and is
different with the hostname of the compute node.

The remote attestation service use the compute node's hostname because
it's the compute node that will run the servers.

Closes-Bug: #1223452

Change-Id: I9a7ce74d595531196804615a8947e253b0bd3f1a