Updated openstack/openstack Project: openstack/keystonemiddleware  89eddf54eea800610407a7995cbedca3498c1382
Replace httpretty with requests-mock

We've had some issue in the past with httpretty breaking compatability
between releases and being a problem to package upstream. requests-mock
is on stackforge and can be tailored to our needs.

Due to issues with py34 and httpretty, this is the preferred fix to
allow middleware to test against py34 (instead of waiting for a fix to
httpretty).

Change-Id: I9fa0f12c732eb187fcb5b0926593cb3624ddc51a