Updated openstack/openstack Project: openstack/keystone  fc2da697e6eaaab0ecfdaa9571d8922d40ebc702
Check for empty string value in REMOTE_USER

The external auth method shouldn't be used if REMOTE_USER is set
to the empty string value. This happens in some setups of Apache
and Shibboleth SP, leading to authentication failures with the
saml2 auth method unless the external auth method is disabled.

Closes-Bug: #1354315
Change-Id: I3cb46ddffa76be3d4526a175257014ca7f1ab94a
Co-Authored-By: Florent Flament <florent.flament-ext@cloudwatt.com>