Updated openstack/openstack Project: openstack/horizon  1a79964e8f94f2803e495591a28efeb504eb4fc6
Make some new strings translator friendly

_usage_summary.html:
blocktrans contains non-simple HTML tags and the range
of blocktrans seems natural. This commit split blocktrans
into the start and end part of the period.

templates/images/properties/_edit.html:
The sensence and the value are concatenated in the code
and translators cannot control the word order.

Change-Id: I3021d67dcb20968b785a8a98ed639a128e3412f2
Closes-Bug: #1354125