Updated openstack/openstack Project: openstack/nova  31f2f379786068c2a0c18271c5caa7424e751eb2
Add unit tests to cells conductor link

Adding unit tests to verify ConductorTaskRPCAPIRedirect
and compute/api.py are properly hooked.  If they become out
of alignment then calls can be potentially ignored.  We can't
blindly error in the case of ignoring as this is used
legitimately in other cases.

Change-Id: Iec02d4a5906a5851c1e9c8258a2d77f908ec65bc
Partial-Bug: 1348642