Updated openstack/openstack Project: openstack/nova  2866b2c26fc1850a8b68bde27ccd6044c40bcc0f
Add hacking check for explicit import of _()

To ensure the right message catalog is used when translating
messages we need to make sure to explicitly import '_' in
any files that use that function.  We cannot count on
unit test to catch cases where the user has forgotten to
import the _() function.

This hacking check ensures that the function has been imported
anywhere that it is used.  Unit tests for the hacking check are
included.

Change-Id: I9d8101916bcb449345d3123617c2ac75776d053e