Updated openstack/openstack Project: openstack/nova  bbc1eb01e134b9c6d5f6f37b3bf3c01afe48387e
Network: interface attach and detach raised confusing exception

When the aforementioned operations failed they raised an exception
that was not coherent - this is due to the fact that the instance object
was passed to the exception. That would print the whole instance object
which is really confusing to a openstack user. The exceptions should has passed
the instance ID and not the instance.

TrivialFix

Change-Id: If3fa89b17210c1db3540cd813157b51e786e1494
Closes-bug: #1335076