Updated openstack/openstack Project: openstack/keystone  89854d3acde9d18c1ffe2d29017e88bc43b37a09
Remove ec2 contrib dependency on token_api

The ec2 contrib extension now utilizes the `validate_token` method
from the `token_provider_api` instead of directly referencing the
token persistence service.

This change is to ensure interactions with the token persistence
are consistent so that it is possible to support toggling
persistence of tokens.

Change-Id: Iad6c13d027fffdb23497f22b29a9ce6d5009f158
bp: non-persistent-tokens