Updated openstack/openstack Project: openstack/horizon  bf3d2721b15a1ec883d56dedd24af54e7ac7b81b
Allow access to HTML final class string

The class_string property does NOT return the final class string.
This is evident in some of the BatchAction classes (which is a subclass).
The only way to get the final class string is through the get_final_attrs
method.

In order to retrieve the final class string, we would need to calculate
the final attributes. This can be cumbersome if all we want is the final
class string. This patch re-factored the code so that the final
class string can be retrieved independently.

Change-Id: I8f56ab407ac6426781cee4b531494d3ef2d41ea1
Closes-Bug: #1347071