Updated openstack/openstack Project: openstack/keystone  3e80e0b005fb045acd83c8879169824f31a125b3
Remove assignment controller dependency on token_api

The assignment controller now uses the `validate_token` method on
the `token_provider_api` instead of referencing `get_token` directly
from the token persistence service.

The user_crud contrib (v2.0) extension now will emit an actual
token to the wire instead of the raw-internal data structure from
the token persistence backend. The intent here was clear by the
fact that the structure was built to look like a v2.0 token,
however, it was retriving the data from the persistence backend
and just updating the token_id then resaving the data before
emitting the raw persistence data (incorrect). The extension
now uses the issue_v2_token method from the token_provider_api.

This change is to ensure interactions with the token persitence
are consistent so that it is possible to support toggling
persistence of tokens.

Change-Id: I22b6b837731566bfd4dc61a47eec2eb0d336d4d1
bp: non-persistent-tokens