Updated openstack/openstack Project: openstack/python-neutronclient  03c6d6a98e7c1c96a3249e69545d100774e0ce98
Provide support for nested objects

With the introduction of the LBaaS v2 API, all objects are capable
of normal CRUD operations. Until now these objects were mostly top-level
constructs, and in then instances where they weren't, they were
manipulated by way of 'actions.'

LBaaS v2, as designed, nests members under pools, and supports all
normal CRUD operations. This means that in order to take action against
a member we need to provide a pool uuid as well as the member uuid. The
python-neutronclient currently does not support this, so this change
makes to possible to do so.

This change adds the capability but does not affect other top-level
object implemenations.

Change-Id: I38c3b78dc077f1f78ebc7e703fe7277a02f53f7a
Partially-implements: blueprint lbaas-api-and-objmodel-improvement