Updated openstack/openstack Project: openstack/tempest  fceab118642fe4c1ffe1fb1bc61af5ec717e6733
Add min_client_version decorator for CLI tests

This adds the min_client_version decorator to be used in the CLI tests.
The decorator is used skip CLI tests where the command being tested
isn't available in the version of the client we're testing against.

The cmd method is pulled out of the ClientTestBase class so it can be
used by the decorator and is renamed so it does not conflict with the
cmd argument passed to it.

To start using the decorator and show how it works, two known CLI tests
are decorated for minimum version checks on their clients.

Part of blueprint minversion-check-for-cli-tests

Change-Id: I8a0ee390048870d450772fa064d0dc2daa2b7036