Updated openstack/openstack Project: openstack/keystone  aa172b831d2ae78f4d52daa95dc9cc81044c477c
Enhance V3 extension class to use resources

RESTful APIs are resource-oriented, and the V3 extensions each
provide a RESTful API, but the current way that extensions map their
routes is based on URLs rather than resources. With this change, the
base V3 extensions class is enhanced to support a resource-oriented
way to map paths to controllers and methods.

bp json-home

Change-Id: Ic82fc9d03b0073a8a1fcec5d03529c84e22f4d69