Updated openstack/openstack Project: openstack/neutron  e2c72495f4cacbafb650fb1e429ea68ba8c07836
Remove unneeded device_owner field from l2pop tuple

The DVR development added this device_owner to the middle of this
tuple during early development because it was thought to be needed.
Over the course of development, it was found to be unnecessary and
much of the code that read it from this value was removed or
obsoleted.  That job went unfinished and so this commit completes it.
This essentially restores the code to what it was before and fixes the
regression that was caused.

Change-Id: Ia901f925883b53e9880dd25688e16e0ffe402bf4
Partial-bug: #1352801